refactor: WorkerFactory를 WorkInjection 내부 클래스로 이동
